Only paste links when pasted text is only url
When the text being pasted contained a URL, we were attempting to convert the target text into a markdown link. However, this was causing some undeseried behavior, because the pasted text was being treated like an exact URL. This updates the RegExp test to look for an exact string of a URL so that we don't match on longer strings.

Borrowed from SO: https://stackoverflow.com/questions/3809401/what-is-a-good-regular-expression-to-match-a-url